get rid of warnings

+++ jsvm/Var.h
@@ -262,7 +262,7 @@ public:
 	}
 
 	inline void removeAllChildren() {
-		for(int i=0; i<children.size(); ++i) {
+		for(size_t i=0; i<children.size(); ++i) {
 			delete children[i];
 		}
 		children.clear();
@@ -348,7 +348,7 @@ public:
 
 	//get child var by index
 	inline BCNode* getChild(int index) {
-		if(index < 0 || index >= children.size())
+		if(index < 0 || index >= (int)children.size())
 			return NULL;
 		return children[index];
 	}
@@ -405,7 +405,7 @@ public:
 				return r;
 		}
 
-		for(int i=0; i<children.size(); ++i) {
+		for(size_t i=0; i<children.size(); ++i) {
 			if(children[i] != NULL && children[i]->name == name)
 				return children[i];
 		}
